requirements.txt
streamlit
tensorflow
matplotlib
numpy
Pillow
gdown
